Refactor: move ray postsolve into RayPostsolver.c
