+++ include/core/Postsolver.h
@@ -24,6 +24,11 @@
 
 #include "PSLP_sol.h"
 #include "Tags.h"
+#include "glbopts.h"
+
+#define COL_NOT_RETRIEVED INF
+#define ROW_NOT_RETRIEVED INF
+#define DUMMY_VALUE -382749
 
 struct u16Vec;
 struct dVec;
@@ -86,6 +91,11 @@ void postsolver_run_primal_infeas_ray(const PostsolveInfo *info, Solution *sol,
                                       const double *y, const double *z);
 void postsolver_run_dual_infeas_ray(const PostsolveInfo *info, Solution *sol,
                                     const double *x);
+
+void retrieve_deleted_row(Solution *sol, int row, double val);
+void retrieve_added_row(Solution *sol, const int *rows, const double *vals);
+void retrieve_added_rows(Solution *sol, int i, const int *rows, const double *vals,
+                         int len, double aik);
 
 /* Saves the information required to retrieve variable xk that was fixed
    to val.
